pyproject.toml
[tool.poetry]
name = "pubmed-papers"
version = "0.1.0"
description = "A tool to fetch research papers from PubMed with company affiliations"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
pandas = "^2.1.0"
biopython = "^1.81"
typer = "^0.9.0"
rich = "^13.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
mypy = "^1.5.1"
isort = "^5.12.0"

[tool.poetry.scripts]
get-papers-list = "pubmed_papers.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 